requirements.txt
altgraph>=0.17.2
Pillow>=9.2.0
Tcl>=0.2
tk>=0.1.0